+++ packages/element/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Bug Fixes
+
+-   `createInterpolateElement`: Return the original string instead of throwing when an unmatched closing tag is present ([#60843](https://github.com/WordPress/gutenberg/issues/60843)).
+
 ## 8.3.0 (2026-07-14)
 
 ## 8.2.0 (2026-07-01)